fix(core): avoid explicit cast and let caller handle it

+++ includes/acl/core/impl/compressed_tracks.impl.h
@@ -124,7 +124,7 @@ namespace acl
 
 	inline bool compressed_tracks::has_trivial_default_values() const { return acl_impl::get_tracks_header(*this).get_has_trivial_default_values(); }
 
-	inline float compressed_tracks::get_default_scale() const { return float(acl_impl::get_tracks_header(*this).get_default_scale()); }
+	inline int32_t compressed_tracks::get_default_scale() const { return acl_impl::get_tracks_header(*this).get_default_scale(); }
 
 	inline sample_looping_policy compressed_tracks::get_looping_policy() const
 	{
